@@ -1,7 +1,7 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk.Web">
 
   <PropertyGroup>
-    <TargetFramework>netcoreapp2.2</TargetFramework>
+    <TargetFramework>netcoreapp3.0</TargetFramework>
     <AssemblyName>Havit.NewProjectTemplate.Web</AssemblyName>
     <RootNamespace>Havit.NewProjectTemplate.Web</RootNamespace>
 	<TypeScriptCompileBlocked>true</TypeScriptCompileBlocked>
@@ -11,13 +11,8 @@
 
   <ItemGroup>
     <PackageReference Include="Havit.AspNetCore.Mvc" Version="2.0.1" />
-    <PackageReference Include="Microsoft.ApplicationInsights.AspNetCore" Version="2.5.1" />
-    <PackageReference Include="Microsoft.AspNetCore" Version="2.2.0" />
-    <PackageReference Include="Microsoft.AspNetCore.Mvc" Version="2.2.0" />
-    <PackageReference Include="Microsoft.AspNetCore.SpaServices" Version="2.2.0" />
-    <PackageReference Include="Microsoft.AspNetCore.StaticFiles" Version="2.2.0" />
-    <PackageReference Include="Microsoft.Extensions.Logging.EventLog" Version="2.2.0" />
-    <PackageReference Include="Microsoft.VisualStudio.Web.BrowserLink" Version="2.2.0" />
+    <PackageReference Include="Microsoft.ApplicationInsights.AspNetCore" Version="2.8.1" />
+	<PackageReference Include="Microsoft.AspNetCore.SpaServices.Extensions" Version="3.0.0" />
 	<PackageReference Include="TypeLite.Lib" Version="1.8.4" NoWarn="NU1701" />
 	<PackageReference Include="Yarn.MSBuild" Version="1.7.0" />
   </ItemGroup>
